==> 4a_OMEGA_global.Rmd <==
---
title: "Vertical velocity (OMEGA) Descriptive Time Series analysis by location"
author: "Shu Wei"
date: ""
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,fig.width=3, fig.height=3)
```

```{r, include = FALSE}
library(tidyverse)
library(fracdiff)
library(ggplot2)
library(ggcorrplot)
library(corrplot)
library(RColorBrewer)
library(forecast)
library(fpp2)
library(ggfortify)
library(dplR)
library(wavethresh)
library(WaveletComp)
library(astsa)
```

Reading data

```{r}
load("D:/UCR/CIMPA/Gauss server/Emuladores/datos/OMEGA_Global.Rdata")
y1 <- OMEGA_Global %>% filter(lat>=30 & lat<=31 & lon>=240 & lon <=241)
y2 <- OMEGA_Global %>% filter(lat>=30 & lat<=31 & lon>=260 & lon <=261)
y3 <- OMEGA_Global %>% filter(lat>=30 & lat<=31 & lon>=291 & lon <=292)
y4 <- OMEGA_Global %>% filter(lat>=41 & lat<=42 & lon>=240 & lon <=241)
y5 <- OMEGA_Global %>% filter(lat>=41 & lat<=42 & lon>=260 & lon <=261)
y6 <- OMEGA_Global %>% filter(lat>=41 & lat<=42 & lon>=291 & lon <=292)
y7 <- OMEGA_Global %>% filter(lat>=51 & lat<=52 & lon>=240 & lon <=241)
y8 <- OMEGA_Global %>% filter(lat>=51 & lat<=52 & lon>=260 & lon <=261)
y9 <- OMEGA_Global %>% filter(lat>=51 & lat<=52 & lon>=291 & lon <=292)

datos <- data.frame(y1[,1:2],y1$OMEGA,y2$OMEGA,y3$OMEGA,y4$OMEGA,y5$OMEGA,y6$OMEGA,y7$OMEGA,y8$OMEGA,y9$OMEGA)
colnames(datos)[3:11]<-c("y1","y2","y3","y4","y5","y6","y7","y8","y9")
names(datos)

datos.ts<-ts(datos[,-c(1,2)])
```


```{r,fig.width=6, fig.height=6}
autoplot(datos.ts)
autoplot(datos.ts, facets = TRUE)
```


# Descriptive: Time Series analysis by location

```{r}
(corr<-cor(datos.ts))
#ggcorrplot(corr)
corrplot(corr, type="upper", order="hclust",
         col=brewer.pal(n=8, name="RdYlBu"))
```

```{r}
datos.ts<-ts(datos)
y1<-datos.ts[,"y1"]
y2<-datos.ts[,"y2"]
y3<-datos.ts[,"y3"]
y4<-datos.ts[,"y4"]
y5<-datos.ts[,"y5"]
y6<-datos.ts[,"y6"]
y7<-datos.ts[,"y7"]
y8<-datos.ts[,"y8"]
y9<-datos.ts[,"y9"]
```

# Autocorrelation:

```{r,fig.width=6, fig.height=6}
par(mfrow=c(3,3))
acf(y1,lag.max=100,ci=0.95)
acf(y2,lag.max=100,ci=0.95)
acf(y3,lag.max=100,ci=0.95)
acf(y4,lag.max=100,ci=0.95)
acf(y5,lag.max=100,ci=0.95)
acf(y6,lag.max=100,ci=0.95)
acf(y7,lag.max=100,ci=0.95)
acf(y8,lag.max=100,ci=0.95)
acf(y9,lag.max=100,ci=0.95)
```


# location 1

```{r,fig.width=6, fig.height=6}
par(mfrow=c(3,3))
ts.plot(y1)
acf(y1,lag.max=100,ci=0.95)
pacf(y1,lag.max=100,ci=0.95)
ts.plot(y4)
acf(y4,lag.max=100,ci=0.95)
pacf(y4,lag.max=100,ci=0.95)
ts.plot(y5)
acf(y5,lag.max=100,ci=0.95)
pacf(y5,lag.max=100,ci=0.95)
```

# Spectral analysis

```{r,fig.width=12, fig.height=6}
# Spectral Analysis
par(mfrow=c(1,2))
raw.spec1 <- spec.pgram(y1, taper = 0)
plot(raw.spec1, log = "no")
raw.spec4 <- spec.pgram(y4, taper = 0)
plot(raw.spec4, log = "no")
raw.spec5 <- spec.pgram(y5, taper = 0)
plot(raw.spec5, log = "no")

ts.plot(datos.ts, col=1:9)
spec = mvspec(datos.ts, spans=c(3,3), taper=.1)

```


# Wavelets analysis

The WaveletCompu library uses Morlet wavelets.

```{r,include=FALSE}
my.data <- data.frame(datos.ts)

my.w1 <- analyze.wavelet(my.data, "y1",loess.span = 0,
                         dt = 1, dj = 1/250,lowerPeriod = 1,
                         upperPeriod = 128,make.pval = TRUE, n.sim = 10)
my.w2 <- analyze.wavelet(my.data, "y2",loess.span = 0,
                         dt = 1, dj = 1/250,lowerPeriod = 1,
                         upperPeriod = 128,make.pval = TRUE, n.sim = 10)
my.w3 <- analyze.wavelet(my.data, "y3",loess.span = 0,
                         dt = 1, dj = 1/250,lowerPeriod = 1,
                         upperPeriod = 128,make.pval = TRUE, n.sim = 10)
my.w4 <- analyze.wavelet(my.data, "y4",loess.span = 0,
                         dt = 1, dj = 1/250,lowerPeriod = 1,
                         upperPeriod = 128,make.pval = TRUE, n.sim = 10)
my.w5 <- analyze.wavelet(my.data, "y5",loess.span = 0,
                         dt = 1, dj = 1/250,lowerPeriod = 1,
                         upperPeriod = 128,make.pval = TRUE, n.sim = 100)
my.w6 <- analyze.wavelet(my.data, "y6",loess.span = 0,
                         dt = 1, dj = 1/250,lowerPeriod = 1,
                         upperPeriod = 128,make.pval = TRUE, n.sim = 10)
my.w7 <- analyze.wavelet(my.data, "y7",loess.span = 0,
                         dt = 1, dj = 1/250,lowerPeriod = 1,
                         upperPeriod = 128,make.pval = TRUE, n.sim = 10)
my.w8 <- analyze.wavelet(my.data, "y8",loess.span = 0,
                         dt = 1, dj = 1/250,lowerPeriod = 1,
                         upperPeriod = 128,make.pval = TRUE, n.sim = 10)
my.w9 <- analyze.wavelet(my.data, "y9",loess.span = 0,
                         dt = 1, dj = 1/250,lowerPeriod = 1,
                         upperPeriod = 128,make.pval = TRUE, n.sim = 10)
```


```{r ,fig.width=5, fig.height=5}
wt.image(my.w1, color.key = "quantile", n.levels = 250,
         legend.params = list(lab = "wavelet power levels", mar = 4.7))
wt.image(my.w2, color.key = "quantile", n.levels = 250,
         legend.params = list(lab = "wavelet power levels", mar = 4.7))
wt.image(my.w3, color.key = "quantile", n.levels = 250,
         legend.params = list(lab = "wavelet power levels", mar = 4.7))
wt.image(my.w4, color.key = "quantile", n.levels = 250,
         legend.params = list(lab = "wavelet power levels", mar = 4.7))
wt.image(my.w5, color.key = "quantile", n.levels = 250,
         legend.params = list(lab = "wavelet power levels", mar = 4.7))
wt.image(my.w6, color.key = "quantile", n.levels = 250,
         legend.params = list(lab = "wavelet power levels", mar = 4.7))
wt.image(my.w7, color.key = "quantile", n.levels = 250,
         legend.params = list(lab = "wavelet power levels", mar = 4.7))
wt.image(my.w8, color.key = "quantile", n.levels = 250,
         legend.params = list(lab = "wavelet power levels", mar = 4.7))
wt.image(my.w9, color.key = "quantile", n.levels = 250,
         legend.params = list(lab = "wavelet power levels", mar = 4.7))
```
